Changes to increase the coverage
